+++ tests/common.py
@@ -1,13 +1,13 @@
-from enum import Enum
+from enum import IntEnum, Enum
 from typing import List
 from datetime import datetime
 
 from pydantic import BaseModel
 
 
-class Order(Enum):
-    asce = True
-    desc = False
+class Order(IntEnum):
+    asce = 1
+    desc = 0
 
 
 class Query(BaseModel):
@@ -32,15 +32,33 @@ class Language(str, Enum):
 class Headers(BaseModel):
     Lang: Language
 
+    class Config:
+        case_sensitive = False
+
+
+class UpperHeaders(BaseModel):
+    LANG: Language
+
+
+class LowerHeaders(BaseModel):
+    lang: Language
+
 
 class Cookies(BaseModel):
-    domain: str
-    path: str
-    value: str
-    expires: datetime
+    pub: str
 
 
 class DemoModel(BaseModel):
     uid: int
     limit: int
     name: str
+
+
+def get_paths(spec):
+    paths = []
+    for path in spec['paths']:
+        if spec['paths'][path]:
+            paths.append(path)
+
+    paths.sort()
+    return paths
